Client FrameWork done

+++ BreadSticks.java
@@ -15,7 +15,7 @@ public class BreadSticks extends Item {
     }
 
     @Override
-    public double cost() {
+    public float cost() {
         return super.getPrice();
     }
